refacto role and group to avoid drawer to close
