+++ opencti-platform/opencti-front/src/private/components/settings/groups/Root.tsx
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
-import React, { FunctionComponent, useMemo } from 'react';
+import React, {FunctionComponent, useMemo, useState} from 'react';
 import { Route, Routes, useParams } from 'react-router-dom';
 import { graphql, PreloadedQuery, usePreloadedQuery, useSubscription } from 'react-relay';
 import { GraphQLSubscriptionConfig } from 'relay-runtime';
@@ -12,10 +12,18 @@ import Group from './Group';
 import { RootGroupsSubscription } from './__generated__/RootGroupsSubscription.graphql';
 import { RootGroupQuery } from './__generated__/RootGroupQuery.graphql';
 import Security from '../../../../utils/Security';
-import { SETTINGS_SETACCESSES } from '../../../../utils/hooks/useGranted';
+import useGranted, {KNOWLEDGE_KNUPDATE_KNDELETE, SETTINGS_SETACCESSES} from '../../../../utils/hooks/useGranted';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { useFormatter } from '../../../../components/i18n';
 import useSensitiveModifications from '../../../../utils/hooks/useSensitiveModifications';
+import Typography from "@mui/material/Typography";
+import PopoverMenu from "../../../../components/PopoverMenu";
+import {Box, styled} from "@mui/material";
+import MenuItem from "@mui/material/MenuItem";
+import GroupDeletionDialog from "@components/settings/groups/GroupDeletionDialog";
+import GroupEdition from "@components/settings/groups/GroupEdition";
+import {useTheme} from "@mui/styles";
+import type {Theme} from "../../../../components/Theme";
 
 const subscription = graphql`
     subscription RootGroupsSubscription($id: ID!) {
@@ -58,11 +66,25 @@ const RootGroupComponent: FunctionComponent<RootGroupComponentProps> = ({ queryR
     [groupId],
   );
   useSubscription(subConfig);
+  const { t_i18n } = useFormatter();
+  const theme = useTheme<Theme>();
+
+  const canDelete = useGranted([KNOWLEDGE_KNUPDATE_KNDELETE]);
+  const [openDelete, setOpenDelete] = useState(false);
+
+  const handleOpenDelete = () => setOpenDelete(true);
+  const handleCloseDelete = () => setOpenDelete(false);
+
   const data = usePreloadedQuery(groupQuery, queryRef);
   const { group } = data;
-  const { t_i18n } = useFormatter();
 
-  const { isSensitive } = useSensitiveModifications('groups', group?.standard_id);
+  const { isAllowed, isSensitive } = useSensitiveModifications('groups', group.standard_id);
+
+  const GroupHeader = styled('div')({
+    display: 'flex',
+    justifyContent: 'space-between',
+    marginBottom: 24,
+  });
 
   return (
     <Security needs={[SETTINGS_SETACCESSES]}>
@@ -78,6 +100,48 @@ const RootGroupComponent: FunctionComponent<RootGroupComponentProps> = ({ queryR
               { label: group.name, current: true },
             ]}
           />
+          <GroupHeader>
+            <div>
+              <Typography
+                variant="h1"
+                gutterBottom={true}
+                style={{ float: 'left' }}
+              >
+                {group.name}
+              </Typography>
+              <div className="clearfix"/>
+            </div>
+            <div style={{ display: 'flex', alignItems: 'center', paddingRight: 200 }}>
+              <div style={{ display: 'flex' }}>
+                <div style={{ marginRight: theme.spacing(0.5) }}>
+                  {canDelete && (
+                    <PopoverMenu>
+                      {({ closeMenu }) => (
+                        <Box>
+                          <MenuItem onClick={() => {
+                            handleOpenDelete();
+                            closeMenu();
+                          }}
+                          >
+                            {t_i18n('Delete')}
+                          </MenuItem>
+                        </Box>
+                      )}
+                    </PopoverMenu>
+                  )}
+                </div>
+                <GroupDeletionDialog
+                  groupId={group.id}
+                  isOpen={openDelete}
+                  handleClose={handleCloseDelete}
+                />
+                <GroupEdition
+                  groupId={group.id}
+                  disabled={!isAllowed && isSensitive}
+                />
+              </div>
+            </div>
+          </GroupHeader>
           <Routes>
             <Route
               path="/"
