+++ opencti-platform/opencti-front/src/private/components/settings/roles/Root.tsx
@@ -1,8 +1,15 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
-import React, { FunctionComponent } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import { Route, Routes, useParams } from 'react-router-dom';
-import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
+import { graphql, PreloadedQuery, useLazyLoadQuery, usePreloadedQuery } from 'react-relay';
+import AccessesMenu from '@components/settings/AccessesMenu';
+import Typography from '@mui/material/Typography';
+import { Box, styled } from '@mui/material';
+import MenuItem from '@mui/material/MenuItem';
+import RoleDeletionDialog from '@components/settings/roles/RoleDeletionDialog';
+import RoleEdition from '@components/settings/roles/RoleEdition';
+import { useTheme } from '@mui/styles';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
@@ -10,11 +17,13 @@ import Role from './Role';
 import { groupsSearchQuery } from '../Groups';
 import { RootRoleQuery } from './__generated__/RootRoleQuery.graphql';
 import { GroupsSearchQuery } from '../__generated__/GroupsSearchQuery.graphql';
-import { SETTINGS_SETACCESSES } from '../../../../utils/hooks/useGranted';
+import useGranted, { SETTINGS_SETACCESSES, KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
 import Security from '../../../../utils/Security';
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import useSensitiveModifications from '../../../../utils/hooks/useSensitiveModifications';
+import PopoverMenu from '../../../../components/PopoverMenu';
+import type { Theme } from '../../../../components/Theme';
 
 const roleQuery = graphql`
   query RootRoleQuery($id: String!) {
@@ -28,6 +37,14 @@ const roleQuery = graphql`
   }
 `;
 
+const roleEditionQuery = graphql`
+  query RootRoleEditionQuery($id: String!) {
+    role(id: $id) {
+      ...RoleEdition_role
+    }
+  }
+`;
+
 interface RootRoleComponentProps {
   queryRef: PreloadedQuery<RootRoleQuery>,
 }
@@ -36,8 +53,18 @@ const RootRoleComponent: FunctionComponent<RootRoleComponentProps> = ({ queryRef
   const data = usePreloadedQuery(roleQuery, queryRef);
   const { role } = data;
   const { t_i18n } = useFormatter();
+  const canDelete = useGranted([KNOWLEDGE_KNUPDATE_KNDELETE]);
+  const theme = useTheme<Theme>();
+  const [openDelete, setOpenDelete] = useState(false);
 
-  const { isSensitive } = useSensitiveModifications('roles', role?.standard_id);
+  const handleOpenDelete = () => setOpenDelete(true);
+  const handleCloseDelete = () => setOpenDelete(false);
+
+  const { isAllowed, isSensitive } = useSensitiveModifications('roles', role?.standard_id);
+  const roleEditionData = useLazyLoadQuery<RoleEditionQuery>(
+    roleEditionQuery,
+    { id: role.id },
+  );
 
   const groupsQueryRef = useQueryLoading<GroupsSearchQuery>(
     groupsSearchQuery,
@@ -48,10 +75,17 @@ const RootRoleComponent: FunctionComponent<RootRoleComponentProps> = ({ queryRef
     },
   );
 
+  const RoleHeader = styled('div')({
+    display: 'flex',
+    justifyContent: 'space-between',
+    marginBottom: 24,
+  });
+
   return (
     <Security needs={[SETTINGS_SETACCESSES]}>
       {role ? (
         <>
+          <AccessesMenu/>
           <Breadcrumbs
             isSensitive={isSensitive}
             elements={[
@@ -61,6 +95,49 @@ const RootRoleComponent: FunctionComponent<RootRoleComponentProps> = ({ queryRef
               { label: role.name, current: true },
             ]}
           />
+          <RoleHeader>
+            <div>
+              <Typography
+                variant="h1"
+                gutterBottom={true}
+                style={{ float: 'left' }}
+              >
+                {role.name}
+              </Typography>
+              <div className="clearfix"/>
+            </div>
+            <div style={{ display: 'flex', alignItems: 'center', paddingRight: 200 }}>
+              <div style={{ display: 'flex' }}>
+                <div style={{ marginRight: theme.spacing(0.5) }}>
+                  {canDelete && (
+                    <PopoverMenu>
+                      {({ closeMenu }) => (
+                        <Box>
+                          <MenuItem onClick={() => {
+                            handleOpenDelete();
+                            closeMenu();
+                          }}
+                          >
+                            {t_i18n('Delete')}
+                          </MenuItem>
+                        </Box>
+                      )}
+                    </PopoverMenu>
+                  )}
+                </div>
+                <RoleDeletionDialog
+                  roleId={role.id}
+                  isOpen={openDelete}
+                  handleClose={handleCloseDelete}
+                />
+                <RoleEdition
+                  roleEditionData={roleEditionData}
+                  disabled={!isAllowed && isSensitive}
+                />
+              </div>
+            </div>
+          </RoleHeader>
+          <div className="clearfix"/>
           <>
             {groupsQueryRef ? (
               <React.Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
